feat: moved NameTerm to another mod
