+++ src/ast/cdl/identifiers/mod.rs
@@ -1,0 +1,3 @@
+mod name_term;
+
+pub use name_term::*;
