+++ src/ast/cdl/mod.rs
@@ -2,8 +2,8 @@ mod actions_and_functions;
 mod entity_term;
 mod field_term;
 mod from_term;
+mod identifiers;
 mod module_term;
-mod name_term;
 mod path_term;
 mod service_term;
 mod types;
@@ -12,8 +12,8 @@ pub use actions_and_functions::*;
 pub use entity_term::*;
 pub use field_term::*;
 pub use from_term::*;
+pub use identifiers::*;
 pub use module_term::*;
-pub use name_term::*;
 pub use path_term::*;
 pub use service_term::*;
 pub use types::*;
